+++ tests/test_mips.c
@@ -17,7 +17,7 @@ struct platform {
 
 static csh handle;
 
-static void print_string_hex(char *comment, unsigned char *str, int len)
+static void print_string_hex(char *comment, unsigned char *str, size_t len)
 {
 	unsigned char *c;
